fix: recommendation screen not visible not route change
